Bump zlib to version 1.2.11. Since no mingw-prebuilt libs are available, build one in he do-test script.

+++ scripts/do-test.sh
@@ -5,18 +5,23 @@ set -e
 uname -a
 ${CC} --version
 sudo apt-add-repository --yes ppa:zoogie/sdl2-snapshots
-sudo apt-get update -qq
-sudo apt-get remove -qq -y mingw32
-sudo apt-get install -qq libsdl2-dev mingw-w64
-mkdir deps
-cd deps
-wget https://www.libsdl.org/release/SDL2-devel-${SDL_VER}-mingw.tar.gz
-tar xfz SDL2-devel-${SDL_VER}-mingw.tar.gz
-wget http://zlib.net/zlib128-dll.zip
-mkdir zlib-${ZLIB_VER}
-cd zlib-${ZLIB_VER}
-unzip ../zlib128-dll.zip
-cd ../..
+sudo apt-get update -qq -y
+sudo apt-get install -qq -y libsdl2-dev
+
+if [ "${CC}" != "clang" ]
+then
+  sudo apt-get remove -qq -y mingw32
+  sudo apt-get install -qq -y mingw-w64
+  mkdir deps
+  cd deps
+  wget https://www.libsdl.org/release/SDL2-devel-${SDL_VER}-mingw.tar.gz
+  tar xfz SDL2-devel-${SDL_VER}-mingw.tar.gz
+  ZLIB_SVER=`echo ${ZLIB_VER} | sed 's|[.]||g'`
+  wget http://zlib.net/zlib${ZLIB_SVER}.zip
+  unzip zlib${ZLIB_SVER}.zip
+  make -C zlib-${ZLIB_VER} PREFIX="${MGW_PREF}-" -f win32/Makefile.gcc
+  cd ..
+fi
 
 make ARCH=LINUX all
 if [ "${CC}" != "clang" ]
